- fix for saving messages; - code cleaning
git-svn-id: http://svn.miranda-ng.org/main/trunk@7318 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/chat/services.cpp
@@ -474,7 +474,7 @@ HWND CreateNewRoom(TContainerData *pContainer, SESSION_INFO *si, BOOL bActivateT
 	newData.iTabID = newItem;
 	newData.iTabImage = newData.item.iImage;
 	newData.pContainer = pContainer;
-	newData.iActivate = (int) bActivateTab;
+	newData.iActivate = bActivateTab;
 	pContainer->iChilds++;
 	newData.bWantPopup = bWantPopup;
 	newData.hdbEvent = (HANDLE)si;
